feat: Support unyank
There are two options to put the unyank ability:
- Option 1: Register a new command like in this CL
- Option 2: Add the unyank to the escape sequence, maybe after the escape_select.

I chose option 1 because the yank function is its own module but the escape
sequence belongs to the tab class. Option 2 is infeasible without touching the
module structure.

I just randomly choose a meaninful and unoccuped key binding for the `unyank`
function. Feel free to modify or remove it from the preset. I personally remap
the yank related functions to dance keys such as ["y" "y"].