feat!: Only return selected when hover is selected
With the cross-directory selection feature, it can be very unexpected
to select items in a parent directory, then go into a child directory
to attempt to open a file, but instead of opening the hovered file,
the selected items are opened instead.

This is especially obvious when using the smart-enter plugin that
combines the enter and open command together.

For the most part, users are most likely going to be hovering over a
selected item if they want to operate on the items that they have
selected, so this change will likely result in a net improvement in
user experience.