refactor: move CLIPBOARD from yazi-core -> yazi-plugin
This needs to be done because we want to make it possible to use the
clipboard in lua plugins.

Before this change, the clipboard was only available in the yazi-core
crate, which depends on yazi-plugin.

The lua plugins are configured in the yazi-plugin crate, but it can't
use anything from yazi-core, because it would create a circular
dependency.