feat(lua): add basic LuaCATS types for internal lua code
This commit adds rudimentary types for the internal lua code. This does
not have a big effect on its own, but it can be a part of a bigger
effort to add types to the codebase.

The types are only usable within yazi itself and are not published for
plugin developers to use (yet).

I only added types for the things that are defined in the lua side of
the codebase. This means most of the useful typing is still missing. I
want to add that separately, as it would be difficult to keep the types
in sync with the actual codebase.