fix(snap): use packages from the archive for dependencies
All of the dependencies previously specified here are available
either in the Ubuntu archive, or as a snap. This commit uses
those packages, which takes the complication out of selecting the
right binaries per architecture, and also from keeping the version
numbers up-to-date.

One exception is the `magick` command which doesn't seem to ship
in the `imagemagick` package on Ubuntu, so we build from source
to ensure it's in-place.