build(snap): remove manual patchelf from cleanup part
Through some experimentation with the arm64 build,
it seems that the magick binary is indeed patched
correctly by snapcraft.

This commit removes an ugly section of the cleanup
part. I've tested the resulting snap, and all
features still work as far as I can tell, despite
linter warnings in the snapcraft build.

This also resolves an issue that was occurring with
the manual patchelf on arm64 systems.