ci: add snap publishing workflow
This commit adds snap publishing and promotion to the CI process.

For nightly releases, upload the built snaps to the 'latest/edge'
channel, enabling users to track nightly builds with:

    snap install yazi --classic --channel latest/edge

For draft releases, upload the built snaps to the 'latest/candidate'
channel.

And finally, when a release is published, promote the snaps that
are in the 'latest/candidate' channel to 'latest/stable'.