fix: `ya pub-to 0` checks if any peer is able to receive the message
Issue
=====

In https://github.com/sxyazi/yazi/issues/2680, `ya pub-to` received the
ability to check if the target peer is able to receive the message that
is being sent. However, if the target peer is `0` (sending to all
peers), the implementation failed.

Solution
========

Check if any peer is able to receive the message when the target peer
is `0`.