fix: Update debian package dependencies
The dependency on `fd-find` causes a removal of `fd` due to a noted conflict and if the user has installed `fd` (i.e. from github or via `deb-get` )
because the renamed `fd-find` in their distro old LTS installation is too old they'll suffer a surprise downgrade or get into a blocked situation.

```
$ apt-cache policy fd-find fd
fd-find:
  Installed: (none)
  Candidate: 8.3.1-1ubuntu0.1
  Version table:
     8.3.1-1ubuntu0.1 500
        500 http://gb.archive.ubuntu.com/ubuntu jammy-updates/universe amd64 Packages
        500 http://gb.archive.ubuntu.com/ubuntu jammy-security/universe amd64 Packages
     8.3.1-1 500
        500 http://gb.archive.ubuntu.com/ubuntu jammy/universe amd64 Packages
fd:
  Installed: 10.3.0
  Candidate: 10.3.0
  Version table:
 *** 10.3.0 100
        100 /var/lib/dpkg/status
```

It might be preferable even to limit the `fd` dependency to >='10.3.0' , but I think the simple alternative would do